+++ projectile.el
@@ -1473,29 +1473,43 @@ With a prefix ARG invalidates the cache first."
   "Return a list of test files for the current project."
   (projectile-test-files (projectile-current-project-files)))
 
-(defvar projectile-rails-rspec '("Gemfile" "app" "lib" "db" "config" "spec"))
-(defvar projectile-rails-test '("Gemfile" "app" "lib" "db" "config" "test"))
-(defvar projectile-symfony '("composer.json" "app" "src" "vendor"))
-(defvar projectile-ruby-rspec '("Gemfile" "lib" "spec"))
-(defvar projectile-ruby-test '("Gemfile" "lib" "test"))
-(defvar projectile-django '("manage.py"))
-(defvar projectile-python-pip '("requirements.txt"))
-(defvar projectile-python-pkg '("setup.py"))
-(defvar projectile-python-tox '("tox.ini"))
-(defvar projectile-scons '("SConstruct"))
-(defvar projectile-maven '("pom.xml"))
-(defvar projectile-gradle '("build.gradle"))
-(defvar projectile-grails '("application.properties" "grails-app"))
-(defvar projectile-lein-test '("project.clj"))
-(defvar projectile-lein-midje '("project.clj" ".midje.clj"))
-(defvar projectile-rebar '("rebar"))
-(defvar projectile-sbt '("build.sbt"))
-(defvar projectile-make '("Makefile"))
-(defvar projectile-grunt '("Gruntfile.js"))
-(defvar projectile-gulp '("gulpfile.js"))
-(defvar projectile-haskell-cabal '("*.cabal"))
-(defvar projectile-rust-cargo '("Cargo.toml"))
-(defvar projectile-r '("DESCRIPTION"))
+(defvar projectile-project-types (make-hash-table)
+  "A hash table holding all project types that are known to Projectile.")
+
+(defun projectile-register-project-type (project-type marker-files compile-command test-command)
+  "Register a project type with projectile.
+
+A project type is defined by PROJECT-TYPE, a set of MARKER-FILES,
+a COMPILE-COMMAND and a TEST-COMMAND."
+  (puthash project-type (list 'marker-files marker-files
+                              'compile-command compile-command
+                              'test-command test-command)
+           projectile-project-types))
+
+(projectile-register-project-type 'rails-rspec '("Gemfile" "app" "lib" "db" "config" "spec") "bundle exec rails server" "bundle exec rspec")
+(projectile-register-project-type 'rails-test '("Gemfile" "app" "lib" "db" "config" "test") "bundle exec railse server" "bundle exec rake test")
+(projectile-register-project-type 'symfony '("composer.json" "app" "src" "vendor") "app/console server:run" "phpunit -c app ")
+(projectile-register-project-type 'ruby-rspec '("Gemfile" "lib" "spec") "bundle exec rake" "bundle exec rspec")
+(projectile-register-project-type 'ruby-test '("Gemfile" "lib" "test") "bundle exec rake" "bundle exec rake test")
+(projectile-register-project-type 'django '("manage.py") "python manage.py runserver" "python manage.py test")
+(projectile-register-project-type 'python-pip '("requirements.txt") "python setup.by build" "python -m unittest discover")
+(projectile-register-project-type 'python-pkg '("setup.py") "python setup.py build" "python -m unittest discover")
+(projectile-register-project-type 'python-tox '("tox.ini") nil "tox")
+(projectile-register-project-type 'scons '("SConstruct") "scons" "scons test")
+(projectile-register-project-type 'maven '("pom.xml") "mvn clean install" "mvn test")
+(projectile-register-project-type 'gradle '("build.gradle") "gradle build" "gradle test")
+(projectile-register-project-type 'grails '("application.properties" "grails-app") "grails package" "grails test-app")
+(projectile-register-project-type 'lein-test '("project.clj") "lein compile" "lein test")
+(projectile-register-project-type 'lein-midje '("project.clj" ".midje.clj") "lein compile" "lein midje")
+(projectile-register-project-type 'rebar '("rebar") "rebar" "rebar eunit")
+(projectile-register-project-type 'sbt '("build.sbt") "sbt compile" "sbt test")
+(projectile-register-project-type 'make '("Makefile") "make" "make test")
+(projectile-register-project-type 'grunt '("Gruntfile.js") "grunt" "grunt test")
+(projectile-register-project-type 'gulp '("gulpfile.js") "gulp" "gulp test")
+(projectile-register-project-type 'haskell-cabal '("*.cabal") "cabal build" "cabal test")
+(projectile-register-project-type 'rust-cargo '("Cargo.toml") "cargo build" "cargo test")
+(projectile-register-project-type 'r '("DESCRIPTION") "R CMD INSTALL ." (concat "R CMD check -o " temporary-file-directory " ."))
+(projectile-register-project-type 'go #'projectile-go "go build ./..." "go test ./...")
 
 (defun projectile-go ()
   "Check if a project contains Go source files."
@@ -1512,35 +1526,19 @@ With a prefix ARG invalidates the cache first."
   "Buffer local var for overriding the auto-detected project type.
 Normally you'd set this from .dir-locals.el.")
 
+(defun projectile-detect-project-type ()
+  (-first (lambda (project-type)
+            (let ((marker (plist-get (gethash project-type projectile-project-types) 'marker-files)))
+              (if (listp marker)
+                  (and (projectile-verify-files marker) project-type)
+                (and (funcall marker) project-type))))
+          (projectile-hash-keys projectile-project-types)))
+
 (defun projectile-project-type ()
   "Determine the project's type based on its structure."
-  (cond
-   (projectile-project-type projectile-project-type)
-   ((projectile-verify-files projectile-rails-rspec) 'rails-rspec)
-   ((projectile-verify-files projectile-rails-test) 'rails-test)
-   ((projectile-verify-files projectile-ruby-rspec) 'ruby-rspec)
-   ((projectile-verify-files projectile-ruby-test) 'ruby-test)
-   ((projectile-verify-files projectile-django) 'django)
-   ((projectile-verify-files projectile-python-tox) 'python-tox)
-   ((projectile-verify-files projectile-python-pip)'python)
-   ((projectile-verify-files projectile-python-pkg) 'python)
-   ((projectile-verify-files projectile-symfony) 'symfony)
-   ((projectile-verify-files projectile-lein-midje) 'lein-midje)
-   ((projectile-verify-files projectile-lein-test) 'lein-test)
-   ((projectile-verify-files projectile-scons) 'scons)
-   ((projectile-verify-files projectile-maven) 'maven)
-   ((projectile-verify-files projectile-gradle) 'gradle)
-   ((projectile-verify-files projectile-grails) 'grails)
-   ((projectile-verify-files projectile-rebar) 'rebar)
-   ((projectile-verify-files projectile-sbt) 'sbt)
-   ((projectile-verify-files projectile-make) 'make)
-   ((projectile-verify-files projectile-gulp) 'gulp)
-   ((projectile-verify-files projectile-grunt) 'grunt)
-   ((projectile-verify-files projectile-haskell-cabal) 'haskell-cabal)
-   ((projectile-verify-files projectile-rust-cargo) 'rust-cargo)
-   ((projectile-verify-files projectile-r) 'r)
-   ((funcall projectile-go-function) 'go)
-   (t 'generic)))
+  (if projectile-project-type
+      projectile-project-type
+    (or (projectile-detect-project-type) 'generic)))
 
 (defun projectile-project-info ()
   "Display info for current project."
@@ -2029,83 +2027,6 @@ For git projects `magit-status-internal' is used if available."
   "Serializes the memory cache to the hard drive."
   (projectile-serialize projectile-projects-cache projectile-cache-file))
 
-(defvar projectile-rails-compile-cmd "bundle exec rails server")
-(defvar projectile-ruby-compile-cmd "bundle exec rake build")
-(defvar projectile-ruby-test-cmd "bundle exec rake test")
-(defvar projectile-ruby-rspec-cmd "bundle exec rspec")
-(defvar projectile-django-compile-cmd "python manage.py runserver")
-(defvar projectile-django-test-cmd "python manage.py test")
-(defvar projectile-python-compile-cmd "python setup.py build")
-(defvar projectile-python-test-cmd "python -m unittest discover")
-(defvar projectile-python-tox-test-cmd "tox")
-(defvar projectile-symfony-compile-cmd "app/console server:run")
-(defvar projectile-symfony-test-cmd "phpunit -c app ")
-(defvar projectile-scons-compile-cmd "scons")
-(defvar projectile-scons-test-cmd "scons test")
-(defvar projectile-maven-compile-cmd "mvn clean install")
-(defvar projectile-maven-test-cmd "mvn test")
-(defvar projectile-gradle-compile-cmd "gradle build")
-(defvar projectile-gradle-test-cmd "gradle test")
-(defvar projectile-grails-compile-cmd "grails package")
-(defvar projectile-grails-test-cmd "grails test-app")
-(defvar projectile-lein-compile-cmd "lein compile")
-(defvar projectile-lein-test-cmd "lein test")
-(defvar projectile-lein-midje-test-cmd "lein midje")
-(defvar projectile-rebar-compile-cmd "rebar")
-(defvar projectile-rebar-test-cmd "rebar eunit")
-(defvar projectile-sbt-compile-cmd "sbt compile")
-(defvar projectile-sbt-test-cmd "sbt test")
-(defvar projectile-make-compile-cmd "make")
-(defvar projectile-make-test-cmd "make test")
-(defvar projectile-grunt-compile-cmd "grunt")
-(defvar projectile-grunt-test-cmd "grunt test")
-(defvar projectile-gulp-compile-cmd "gulp")
-(defvar projectile-gulp-test-cmd "gulp test")
-(defvar projectile-go-compile-cmd "go build ./...")
-(defvar projectile-go-test-cmd "go test ./...")
-(defvar projectile-haskell-cabal-compile-cmd "cabal build")
-(defvar projectile-haskell-cabal-test-cmd "cabal test")
-(defvar projectile-rust-cargo-compile-cmd "cargo build")
-(defvar projectile-rust-cargo-test-cmd "cargo test")
-(defvar projectile-r-compile-cmd "R CMD INSTALL .")
-(defvar projectile-r-test-cmd (concat "R CMD check -o "
-                                      temporary-file-directory " ."))
-
-(--each '(projectile-rails-compile-cmd
-          projectile-ruby-compile-cmd
-          projectile-ruby-test-cmd
-          projectile-ruby-rspec-cmd
-          projectile-django-compile-cmd
-          projectile-django-test-cmd
-          projectile-python-compile-cmd
-          projectile-python-test-cmd
-          projectile-python-tox-test-cmd
-          projectile-symfony-compile-cmd
-          projectile-symfony-test-cmd
-          projectile-scons-compile-cmd
-          projectile-scons-test-cmd
-          projectile-maven-compile-cmd
-          projectile-maven-test-cmd
-          projectile-lein-compile-cmd
-          projectile-lein-test-cmd
-          projectile-lein-midje-test-cmd
-          projectile-rebar-compile-cmd
-          projectile-rebar-test-cmd
-          projectile-sbt-compile-cmd
-          projectile-sbt-test-cmd
-          projectile-make-compile-cmd
-          projectile-make-test-cmd
-          projectile-grunt-compile-cmd
-          projectile-grunt-test-cmd
-          projectile-haskell-cabal-compile-cmd
-          projectile-haskell-cabal-test-cmd
-          projectile-rust-cargo-compile-cmd
-          projectile-rust-cargo-test-cmd
-          projectile-r-compile-cmd
-          projectile-r-test-cmd)
-  (put it 'safe-local-variable #'stringp))
-
-
 (defvar projectile-compilation-cmd-map
   (make-hash-table :test 'equal)
   "A mapping between projects and the last compilation command used on them.")
@@ -2115,53 +2036,11 @@ For git projects `magit-status-internal' is used if available."
 
 (defun projectile-default-compilation-command (project-type)
   "Retrieve default compilation command for PROJECT-TYPE."
-  (cond
-   ((member project-type '(rails-rspec rails-test)) projectile-rails-compile-cmd)
-   ((member project-type '(ruby-rspec ruby-test)) projectile-ruby-compile-cmd)
-   ((eq project-type 'django) projectile-django-compile-cmd)
-   ((eq project-type 'python) projectile-python-compile-cmd)
-   ((eq project-type 'symfony) projectile-symfony-compile-cmd)
-   ((member project-type '(lein-test lein-midje)) projectile-lein-compile-cmd)
-   ((eq project-type 'make) projectile-make-compile-cmd)
-   ((eq project-type 'rebar) projectile-rebar-compile-cmd)
-   ((eq project-type 'scons) projectile-scons-compile-cmd)
-   ((eq project-type 'maven) projectile-maven-compile-cmd)
-   ((eq project-type 'gradle) projectile-gradle-compile-cmd)
-   ((eq project-type 'grails) projectile-grails-compile-cmd)
-   ((eq project-type 'sbt) projectile-sbt-compile-cmd)
-   ((eq project-type 'grunt) projectile-grunt-compile-cmd)
-   ((eq project-type 'gulp) projectile-gulp-compile-cmd)
-   ((eq project-type 'go) projectile-go-compile-cmd)
-   ((eq project-type 'haskell-cabal) projectile-haskell-cabal-compile-cmd)
-   ((eq project-type 'rust-cargo) projectile-rust-cargo-compile-cmd)
-   ((eq project-type 'r) projectile-r-compile-cmd)
-   (t projectile-make-compile-cmd)))
+  (plist-get (gethash project-type projectile-project-types) 'compile-command))
 
 (defun projectile-default-test-command (project-type)
   "Retrieve default test command for PROJECT-TYPE."
-  (cond
-   ((member project-type '(rails-rspec ruby-rspec)) projectile-ruby-rspec-cmd)
-   ((member project-type '(rails-test ruby-test)) projectile-ruby-test-cmd)
-   ((eq project-type 'django) projectile-django-test-cmd)
-   ((eq project-type 'python-tox) projectile-python-tox-test-cmd)
-   ((eq project-type 'python) projectile-python-test-cmd)
-   ((eq project-type 'symfony) projectile-symfony-test-cmd)
-   ((eq project-type 'lein-test) projectile-lein-test-cmd)
-   ((eq project-type 'lein-midje) projectile-lein-midje-test-cmd)
-   ((eq project-type 'make) projectile-make-test-cmd)
-   ((eq project-type 'rebar) projectile-rebar-test-cmd)
-   ((eq project-type 'scons) projectile-scons-test-cmd)
-   ((eq project-type 'maven) projectile-maven-test-cmd)
-   ((eq project-type 'gradle) projectile-gradle-test-cmd)
-   ((eq project-type 'grails) projectile-grails-test-cmd)
-   ((eq project-type 'sbt) projectile-sbt-test-cmd)
-   ((eq project-type 'grunt) projectile-grunt-test-cmd)
-   ((eq project-type 'gulp) projectile-gulp-test-cmd)
-   ((eq project-type 'go) projectile-go-test-cmd)
-   ((eq project-type 'haskell-cabal) projectile-haskell-cabal-test-cmd)
-   ((eq project-type 'rust-cargo) projectile-rust-cargo-test-cmd)
-   ((eq project-type 'r) projectile-r-test-cmd)
-   (t projectile-make-test-cmd)))
+  (plist-get (gethash project-type projectile-project-types) 'test-command))
 
 (defun projectile-compilation-command (project)
   "Retrieve the compilation command for PROJECT."