+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add support for projects using Midje.
 * Add the ability to create missing tests automatically (controlled via the `projectile-create-missing-test-files` defcustom).
 * Add the ability to dynamically decide if a project should be added to `projectile-known-projects` (via new `projectile-ignored-project-function` defcustom).
+* Add the ability to register new project types dynamically with `projectile-register-project-type`.
 
 ### Changes